abstracted out butterfly networks
